skill/requirements.txt
futu-api>=10.0.6008
pandas>=2.0.0
